+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.2.0 (2023-04-27)
+
+### Feat
+
+- defined error codes for what linter will cover
+- added import rules module and parse utils and error codes (#9)
+- added rules checker class to process the results of parsing through file (#8)
+- add custom import rules vistor class for traversing imports (#6)
+
 ## v0.1.0 (2023-04-26)
 
 ### Feat